refactor: refactoring project structure
