+++ utils/currency_utils.ts
@@ -1,0 +1,4 @@
+export const formatCurrency = (value: number): string => {
+    return new Intl.NumberFormat('pt-BR', { style: 'currency', currency: 'BRL' }).format(value);
+  };
+
